Update deploy script with helpful comment if no parameter is given

+++ deploy.sh
@@ -1,6 +1,11 @@
 #!/bin/sh
 
 # Helpful for syncing files from generated Hakyll site
-# Example: ./deploy ~/projects/savannamodel-hakyll
+
+if [ -z "$1" ]; then
+    echo "Provide a Hakyll site directory to deploy from."
+    echo "Example: ./deploy ~/projects/savannamodel-hakyll"
+    exit 1
+fi
 
 rsync -av "$1/_site/" docs/ --delete
